+++ index.js
@@ -1,135 +1,64 @@
+'use strict'
 
-
-/*
-const MongoClient = require('mongodb').MongoClient;
-const uri = "mongodb+srv://karina2004liz:ex300172004@firstdatabase-t9rak.gcp.mongodb.net/test?retryWrites=true&w=majority";
-const client = new MongoClient(uri, { useNewUrlParser: true });
-client.connect(err => {
-  const collection = client.db("test").collection("devices");
-  // perform actions on the collection object
-  client.close();
-});
-
-*/
-
-/*
-
-
-const express = require('express');
-
-
-
- const app = express();
-
-app.listen(3001, function() {
-    console.log('listening on 3001')
-  })
-
-  app.get('/products', (req, res) => {
-    res.send(gato)
-  })
-
-  app.get('/', (req, res) => {
-    res.send("Utilice /products para checar existencias")
-  })
-
-
-  var gato = {
-
-    michi :{
-        name : "michi",
-        age : "5years"  
-
-    },
-    michi2 :{
-        name : "michi",
-        age : "5years"  
-
-    },
-
-    michi3 :{
-        name : "michi",
-        age : "5years"  
-
-    }
-
-    ,    michi4 :{
-        name : "michi",
-        age : "5years"  
-
-    },
- 
-
-  }
-
-
-*/
-
+const express = require('express')
+const bodyParser = require('body-parser')
+const mongoose = require('mongoose')
+const app = express()
 const port = process.env.PORT || 3000
+const cors = require('cors');
+const productsFunctions = require('./src/routes/product.js')
+const ordersFunctions = require('./src/routes/order.js')
 
-var cors = require("cors");
-var bodyParser = require("body-parser");
-var express = require("express");
-var app = express();
-var mongose = require("mongoose");
-
-
-
-app.use(cors());
-//app.use(express.json());
-
+app.use(cors())
+app.use(bodyParser.urlencoded({ extended: false }))
+app.use(bodyParser.json())
 
 app.listen(port, err => {
-    if (err) throw err
-    console.log(`> Ready On Server http://localhost:${port}`)
+  if (err) throw err
+  console.log(`> Ready On Server http://localhost:${port}`)
 });
 
-app.get("/get", (req, res, next) => {
-    res.json({
-        "version": process.env.VERSION
-    });
-});
+app.get('/', (req, res)=>{
+  res.send("Hola")
+})
+app.get('/api/products', productsFunctions.getProducts)
+app.get('/api/products/:productId', productsFunctions.getProductId)
+app.post('/api/products', productsFunctions.postProduct)
+app.put('/api/products/:productId', productsFunctions.putProduct)
+app.delete('/api/products/:productId', productsFunctions.deleteProduct)
 
-app.post('/post', function(request, response) {
-    response.send(request.body);
-});
+app.get('/api/orders', ordersFunctions.getOrders)
+app.get('/api/orders/:orderId', ordersFunctions.getOrderId)
+app.post('/api/orders', ordersFunctions.postOrder)
+app.put('/api/orders/:orderId', ordersFunctions.putOrder)
+app.delete('/api/orders/:ordersId', ordersFunctions.deleteOrder)
 
-
-app.get("/products", (req,res)=>{
-  res.send(products)
+mongoose.connect('mongodb+srv://karina2004liz:<password>@firstdatabase-t9rak.gcp.mongodb.net/test?retryWrites=true&w=majority', (err, res) => {
+  if (err) {
+    return console.log(`Error to connect with database: ${err}`)
+  }
 })
 
-var products = {
+let products = {
 
- products: [
+  products: [
 
-    {name : "coffee",
-    price: 15,
-    id: 1,
+    {
+      name: "coffee",
+      price: 15,
+      id: 1,
     },
     {
-    name: "tea",
-    price: 10,
-    id: 2,
+      name: "tea",
+      price: 10,
+      id: 2,
     },
-    {name: "chocolate",
-    price: 20,
-    id: 3
-
-    },
-    {name: "lechera",
-    price: 20,
-    id: 4
+    {
+      name: "chocolate",
+      price: 20,
+      id: 3
 
     }
-    ,
-    {name: "Galleta",
-    price: 20,
-    id: 5
 
-    }
   ]
 }
-
-
-
