+++ now.json
@@ -1,6 +1,6 @@
 {
     "version": 2,
-    "builds": [{ "src": "index.js", "use": "@now/node-server" }],
+    "builds": [{ "src": "/index.js", "use": "@now/node-server" }],
     "routes": [
         {
             "headers": {
@@ -23,7 +23,7 @@
         "methods": ["POST"]
       },
       {
-        "src": "/products",
+        "src": "/api/products",
         "dest": "/index.js",
         "methods": ["GET"]
       }
